Move header-only c10/util string/filesystem utils to torch/headeronly
Migrate three already-header-only c10/util headers into torch/headeronly/util/,
leaving the old c10 paths as one-line forwarders (fully backward compatible, no
consumer include changes):

  string_utils.h, FileSystem.h, string_view.h

Part of the wave moving c10 headers that need no .cpp/libtorch linking into the
header-only library. string_utils (deprecated std re-exports) and FileSystem
(the c10::filesystem namespace alias) are flipped/re-exported into
torch::headeronly.

string_view is deliberately NOT flipped: basic_string_view stays defined in
namespace c10 and is only re-exported up into torch::headeronly. c10/util/
TypeIndex.h parses __PRETTY_FUNCTION__ and hardcodes the printed return type as
c10::basic_string_view<char>; defining the type canonically in torch::headeronly
would change the compiler-printed name and break TypeIndex's static_asserts.

The string_view diff is large only because moving a 643-line header with a
forwarder shows the old content as deleted and the new content as added (git
cannot render it as a rename when the old path becomes a forwarder); the net
logical change is the namespace re-export plus the forwarder.

Test Plan:

Full build (confirms all libtorch consumers still compile, including the
__PRETTY_FUNCTION__-sensitive c10/util/TypeIndex.h):

```
pip install -e . --no-build-isolation
```

Header-only ABI check (compiles + runs without linking libtorch):

```
build/bin/test_aoti_abi_check --gtest_filter='TestStringView*:TestStringUtils*:TestFileSystem*'
```

Lint:

```
lintrunner -a <changed files>
```

This PR was authored with the assistance of Claude (Claude Code).

[ghstack-poisoned]
